# PERF: avoid unnecessary recoding in CategoricalIndex._create_categorical