@gfyoung 2018-06-16
@topper-123 : Nice!  Can you post `asv` benchmark timings (to make sure there were no regressions introduced) if you have a chance?