@topper-123 2018-06-16
Hmm, there seem to be introduced some performance regressions by this PR that I don't quite understand. I will close this PR until I've figured it out, so as not to clutter PR quoue...